requirements.txt
django
djangorestframework
djangorestframework-simplejwt
djangorestframework-filters==1.0.0.dev2
django-cors-headers